Rename: JsonResponseUtil 리팩토링
위치 수정
