+++ src/main/java/com/chapter1/blueprint/security/filter/JwtAuthenticationFilter.java
@@ -1,7 +1,7 @@
 package com.chapter1.blueprint.security.filter;
 
+import com.chapter1.blueprint.exception.util.JsonResponseUtil;
 import com.chapter1.blueprint.security.util.JwtProcessor;
-import com.chapter1.blueprint.util.JsonResponseUtil;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
